Food spawn and eating
The food spawing and eating of the artros has been implemented succesfully.
The hungry artros go blue until they eat.
The search (for example, for food) right now is based on quadrants, which has been really challenging to implement. Right now artros can only see food in their quadrant, which is not correct (there may be food in the adjacent quadrant that is within its vision and it cannot see). Looking for a better way to do the search so that the code doesn't have to look up the entireeeeeee list (of food, for example) and I don't have to use quadrants as well