Export results
Results are being exported to a .txt file in the same folder as the app, as I wanted. The file is generated automatically and it registers the population of artros, sampling every 10 times they act (in other words, for every 100 rounds, only 10 rounds have their population recorded). This is to be able to create nice light graphs that can be easily handled.

Next step is to allow for custom input through a Json and that will conclude the first version!